pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "route53-ddns"
version = "0.1.0"
description = "Keep Route53 A records in sync with public IPv4; status UI and companion TXT timestamps."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "apprise>=1.9",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "jinja2>=3.1",
  "boto3>=1.35",
  "httpx>=0.27",
  "pydantic>=2.10",
  "pydantic-settings>=2.6",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "respx>=0.21",
]

[project.scripts]
route53-ddns = "route53_ddns.main:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
